+++ packages/react/src/experimental/Navigation/ApplicationFrame/index.stories.tsx
@@ -1,6 +1,7 @@
 import { Lightbulb } from "@/icons/app"
 import ExternalLink from "@/icons/app/ExternalLink"
-import { useAiChat } from "@/experimental/AiChat/providers/AiChatStateProvider"
+import { useAiChat } from "@/ai/AiChat/providers/AiChatStateProvider"
+import PalmTree from "@/icons/app/PalmTree"
 import type { Meta, StoryObj } from "@storybook/react-vite"
 import { ComponentProps, useEffect, useRef } from "react"
 import { expect, within } from "storybook/test"
@@ -31,6 +32,18 @@ const meta: Meta<typeof ApplicationFrame> = {
           message: "Share feedback",
           prompt:
             "Share feedback and help shape One with your feedback in the next message (optional)",
+        },
+        {
+          icon: PalmTree,
+          message: "Create a new task",
+          prompt:
+            "Create a new task and help shape One with your feedback in the next message (optional)",
+        },
+        {
+          icon: Lightbulb,
+          message: "Create a new project",
+          prompt:
+            "Create a new project and help shape One with your feedback in the next message (optional)",
         },
       ],
     },