+++ packages/react/src/experimental/OneDataCollection/__stories__/mockData.tsx
@@ -271,7 +271,6 @@ export const getMockVisualizations = (options?: {
                   },
             id: "name",
             sorting: options?.table?.noSorting ? undefined : "name",
-            hidden: options?.table?.allowColumnHiding ? true : undefined,
             order: options?.table?.allowColumnReordering ? 3 : undefined,
           },
           {
@@ -301,12 +300,14 @@ export const getMockVisualizations = (options?: {
             render: (item) => item.email,
             sorting: options?.table?.noSorting ? undefined : "email",
             order: options?.table?.allowColumnReordering ? 1 : undefined,
+            hidden: options?.table?.allowColumnHiding ? true : undefined,
           },
           {
             id: "role2",
             label: "Role 2",
             render: (item) => item.role,
             sorting: options?.table?.noSorting ? undefined : "role",
+            hidden: options?.table?.allowColumnHiding ? true : undefined,
           },
           {
             id: "department2",
@@ -314,6 +315,7 @@ export const getMockVisualizations = (options?: {
             render: (item) => item.department,
             sorting: options?.table?.noSorting ? undefined : "department",
             order: options?.table?.allowColumnReordering ? 10 : undefined,
+            hidden: options?.table?.allowColumnHiding ? true : undefined,
           },
           {
             id: "email3",
@@ -372,7 +374,15 @@ export const getMockVisualizations = (options?: {
           },
         ],
       },
-    },
+    } as Visualization<
+      MockUser,
+      FiltersType,
+      typeof sortings,
+      SummariesDefinition,
+      ItemActionsDefinition<MockUser>,
+      NavigationFiltersDefinition,
+      GroupingDefinition<MockUser>
+    >,
     card: {
       type: "card",
       options: {
@@ -567,7 +577,7 @@ export const getMockVisualizations = (options?: {
             render: (item) => ({
               type: "dotTag",
               value: {
-                color: "yellow",
+                color: "yellow" as const,
                 label: item.department,
               },
             }),
@@ -686,7 +696,7 @@ export const getMockVisualizations = (options?: {
           : undefined,
       },
     },
-  })
+  }) as const
 // Example of using the object-based approach (recommended)
 export const sortings = {
   name: {